[Keymap] Fix bootmagic compilation issue with Drashna keymaps (#18223)
